Allow for a cluster service classes dependencies field
Currently we unmarshal the cluster service class JSON response into
a map[string]string. However, cluster service classes can also have
a dependencies field which is an array of strings. This causes every
provision to fail as the unmarshalling fails.

This change adds a ServiceClass struct instead of the map[string]string.