🔒 Require approval for PRs from untrusted sources
This initial step should not have any impact on PRs from people who
are members of the aerogear org, or for non-PR builds from this
Jenkinsfile.

For contributors who are not members of the aerogear org, PRs will
hang at this step, waiting for someone to approve. Only people who can
authenticate to the Jenkins instance will be able to approve --
currently that means 'members of the fheng org'.

This is intended as a safety measure to help prevent potential attacks
on our build infrastructure, by executing arbitrary code there through
pull requests.