Use environment variable to allow invalid/self-signed certificates
Currently we allow all invalid/self-signed certificates. We need an
option to allow this to be disallowed. This comes in the form of an
environment variable named AEROGEAR_ALLOW_INVALID_CERTS.

Now, by default, the CLI will not allow any invalid certificates to
be used. Only when the environment variable is set to 'true' will
they be allowed.